publish script
