+++ package.json
@@ -1,12 +1,10 @@
 {
   "version": "0.0.17",
+  "private": true,
   "scripts": {
     "test": "deno -A npm:vitest run --testTimeout 0",
     "format": "prettier --write .",
-    "check": "tsc",
-    "update": "deno -A scripts/update.ts",
-    "build": "deno -A cli/build.ts",
-    "publish": "deno task build && cd dist && npm publish"
+    "check": "tsc"
   },
   "workspaces": [
     "./jsgrad",
